[chore] move all target-allocator code to internal
