+++ cmd/otel-allocator/internal/allocation/least_weighted.go
@@ -4,7 +4,7 @@
 package allocation
 
 import (
-	"github.com/open-telemetry/opentelemetry-operator/cmd/otel-allocator/target"
+	"github.com/open-telemetry/opentelemetry-operator/cmd/otel-allocator/internal/target"
 )
 
 const leastWeightedStrategyName = "least-weighted"
