+++ cmd/otel-allocator/internal/allocation/per_node.go
@@ -6,7 +6,7 @@ package allocation
 import (
 	"fmt"
 
-	"github.com/open-telemetry/opentelemetry-operator/cmd/otel-allocator/target"
+	"github.com/open-telemetry/opentelemetry-operator/cmd/otel-allocator/internal/target"
 )
 
 const perNodeStrategyName = "per-node"
