+++ cmd/otel-allocator/internal/config/testdata/config_test.yaml
