+++ cmd/otel-allocator/internal/config/testdata/file_sd_test.json
