+++ cmd/otel-allocator/internal/config/testdata/no_config.yaml
