+++ cmd/otel-allocator/internal/config/testdata/pod_service_selector_expressions_test.yaml
