+++ cmd/otel-allocator/internal/config/testdata/pod_service_selector_test.yaml
