+++ cmd/otel-allocator/internal/diff/diff_test.go
