+++ cmd/otel-allocator/internal/server/mocks_test.go
@@ -4,8 +4,8 @@
 package server
 
 import (
-	"github.com/open-telemetry/opentelemetry-operator/cmd/otel-allocator/allocation"
-	"github.com/open-telemetry/opentelemetry-operator/cmd/otel-allocator/target"
+	"github.com/open-telemetry/opentelemetry-operator/cmd/otel-allocator/internal/allocation"
+	"github.com/open-telemetry/opentelemetry-operator/cmd/otel-allocator/internal/target"
 )
 
 var _ allocation.Allocator = &mockAllocator{}
