+++ cmd/otel-allocator/internal/server/testdata/prom-config-all-actions.yaml
