+++ cmd/otel-allocator/internal/server/testdata/prom-config-test.yaml
