+++ cmd/otel-allocator/internal/server/testdata/prom-no-config.yaml
